requirements.txt
groq>=0.4.0
pandas>=1.5.0
python-dotenv>=0.21.0
pytest>=7.0.0
